requirements.txt

discord.py==1.6.0
errbot==6.1.8
python-dotenv==0.18.0
